Fix stuck shift key bug
If you press LAYER then H (ampersand), then releases LAYER
before H, the shift key would stay stuck down.
This happened because we see ampersand (shift + 7) pressed
but, because LAYER is released first, we see H released.

The fix for MODKEYs is to release the modifier immediately,
while we still know which layer it is from.

The downside of this fix is that MODKEYs can't autorepeat.